requirements.txt
aiohttp
ccxt>=4.0.0
numpy
pandas
pydantic>=2.0.0
pydantic-settings
pymongo
pytest
pytest-asyncio
pytest-mock
python-dotenv
python-logging-loki
python-telegram-bot
redis
requests
structlog
ta
websockets
PyYAML
pyarrow
fastapi
uvicorn
httpx
